+++ src/main/java/com/commercetools/sync/channels/helpers/ChannelCustomActionBuilder.java
@@ -12,25 +12,30 @@ import javax.annotation.Nullable;
 import java.util.Map;
 
 
-public class ChannelCustomActionBuilder extends GenericCustomActionBuilder<Channel> {
+public class ChannelCustomActionBuilder implements GenericCustomActionBuilder<Channel> {
 
     @Nonnull
     @Override
-    public UpdateAction<Channel> buildRemoveCustomTypeAction() {
+    public UpdateAction<Channel> buildRemoveCustomTypeAction(@Nullable final Integer variantId,
+                                                             @Nullable final String objectId) {
         return SetCustomType.ofRemoveType();
     }
 
     @Nonnull
     @Override
 
-    public UpdateAction<Channel> buildSetCustomTypeAction(@Nullable final String customTypeId,
+    public UpdateAction<Channel> buildSetCustomTypeAction(@Nullable final Integer variantId,
+                                                          @Nullable final String objectId,
+                                                          @Nonnull final String customTypeId,
                                                           @Nullable final Map<String, JsonNode> customFieldsJsonMap) {
         return SetCustomType.ofTypeIdAndJson(customTypeId, customFieldsJsonMap);
     }
 
     @Nonnull
     @Override
-    public UpdateAction<Channel> buildSetCustomFieldAction(@Nullable final String customFieldName,
+    public UpdateAction<Channel> buildSetCustomFieldAction(@Nullable final Integer variantId,
+                                                           @Nullable final String objectId,
+                                                           @Nullable final String customFieldName,
                                                            @Nullable final JsonNode customFieldValue) {
         return SetCustomField.ofJson(customFieldName, customFieldValue);
     }