+++ src/main/java/com/commercetools/sync/commons/utils/SyncUtils.java
@@ -81,4 +81,7 @@ public final class SyncUtils {
         }
         return reference;
     }
+
+    private SyncUtils() {
+    }
 }